% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphTools.R
\name{DiceRollHistogram}
\alias{DiceRollHistogram}
\title{A simple histogram function unsing percentage values.}
\usage{
DiceRollHistogram(data, bars)
}
\arguments{
\item{data}{A vector, the data to be graphed.}

\item{bars}{The amount of bars to show. (Default: 10)}
}
\value{
A histogram with percentage values.
}
\description{
A simple histogram function unsing percentage values.
}
\examples{
DiceRollHistogram(dice.roll.vector, bars = length(dice.roll.vector))
DiceRollHistogram(some.vector, bars = 5)
}
