% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTools.R
\name{GetOddsXdYLowerThan}
\alias{GetOddsXdYLowerThan}
\title{Get the odds of rolling XdY dice and not beating a given score.}
\usage{
GetOddsXdYLowerThan(X, odds, Y = 10, theoretical = FALSE, Z = 10000)
}
\arguments{
\item{X}{An integer, the amount of dice to be rolled.}

\item{Y}{An integer, the number of sides the die has. (Default: 10)}

\item{theoretical}{Boolean, whether or not to use an ideal approach. Warning: An ideal approach can be very resource intensive for values of X and/or Y larger than 8. (Default: FALSE)}

\item{score.odda}{An integer, the score not to beat.}

\item{dice.rolls}{An Integer, the amount of dice rolls to be simulated in case of running an empyrical simulation. (Default: 10000)}
}
\value{
An integer with the probability of the score given being achieved or beaten expressed as a percentage.
}
\description{
Get the odds of rolling XdY dice and not beating a given score.
}
\examples{
GetOddsXdYLowerThan(5, score.odds = 7)
GetOddsXdYLowerThan(X = 10, Y = 6, score.odds = 50)
}

