% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTools.R
\name{GetScoreDistribution}
\alias{GetScoreDistribution}
\title{Get the distribution for a total score after rolling XdY dice and summing the results.}
\usage{
GetScoreDistribution(X, Y = 10, dice.rolls = 10000, plot.histogram = TRUE,
  theoretical = FALSE)
}
\arguments{
\item{X}{An integer, the amount of dice to be rolled.}

\item{Y}{An ingeger, the number of sides the die has. (Default: 10)}

\item{dice.rolls}{An Integer, the amount of dice rolls to be simulated in case of running an empyrical simulation. (Default: 10000)}

\item{plot.histogram}{Boolean, whether or not to plot a histogram with the data obtained}

\item{theoretical}{Boolean, whether or not to use an ideal approach. Warning: An ideal approach can be very resource intensive for values of X and/or Y larger than 8. (Default: FALSE)}
}
\value{
A list with either simulated or an exhaustive account of dice roll combinations scores.
}
\description{
Get the distribution for a total score after rolling XdY dice and summing the results.
}
\examples{
GetScoreDistribution(5)
GetScoreDistribution(X = 10, Y = 6, plot.histogram = FALSE)
}
