% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessesTools.R
\name{GetSuccessesDistribution}
\alias{GetSuccessesDistribution}
\title{Get the distribution for a success/failure roll (White Wolf Style).}
\usage{
GetSuccessesDistribution(X, Y = 10, success = 6, fail = 1,
  dice.rolls = 10000, theoretical = FALSE, plot.histogram = TRUE)
}
\arguments{
\item{X}{An integer, the amount of dice to be rolled.}

\item{Y}{An integer, the number of sides the die has. (Default: 10)}

\item{success}{An Integer, the minimum amount to be rolled to be considered a success. (Default: 6)}

\item{fail}{An Integer, the maximum number than if rolled, gets substracted from the success count. (Default: 1)}

\item{dice.rolls}{An Integer, the amount of dice rolls to be simulated in case of running an empyrical simulation. (Default: 10000)}

\item{theoretical}{Boolean, whether or not to use an ideal approach. Warning: An ideal approach can be very resource intensive for values of X and/or Y larger than 8. (Default: FALSE)}

\item{plot.histogram}{Boolean, whether or not to plot a histogram with the data obtained}
}
\value{
A list with either simulated or an exhaustive account of dice roll combinations success counts.
}
\description{
Get the distribution for a success/failure roll (White Wolf Style).
}
\examples{
GetSuccessesDistribution(5)
GetSuccessesDistribution(X = 10, Y = 6, plot.histogram = FALSE)
}

