% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreTools.R
\name{GetTheoreticalScores}
\alias{GetTheoreticalScores}
\title{Get a vector with all possible scores after rolling XdY.  Warning: An ideal approach can be very resource intensive for values of X and/or Y larger than 8.}
\usage{
GetTheoreticalScores(X, Y)
}
\arguments{
\item{X}{An integer, the amount of dice to be rolled.}

\item{Y}{An integer, the number of sides the die has. (Default: 10)}

\item{dice.rolls}{An Integer, the amount of dice rolls to be simulated.}
}
\value{
A Vector with all possbile scores for XdY
}
\description{
Get a vector with all possible scores after rolling XdY.  Warning: An ideal approach can be very resource intensive for values of X and/or Y larger than 8.
}
\examples{
GetTheoreticalScores(5, 10)
GetTheoreticalScores(X = 8, Y = 10)
}

